[refactor] : rename User To MyPage
