+++ presenter/src/main/java/com/mashup/twotoo/presenter/mypage/navigation/UserNavigation.kt
@@ -1,10 +1,10 @@
-package com.mashup.twotoo.presenter.user.navigation
+package com.mashup.twotoo.presenter.mypage.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.mashup.twotoo.presenter.user.UserRoute
+import com.mashup.twotoo.presenter.mypage.MyPageRoute
 
 const val UserNavigationRoute = "user_route"
 
@@ -14,6 +14,6 @@ fun NavController.navigateToUser(navOptions: NavOptions? = null) {
 
 fun NavGraphBuilder.userGraph() {
     composable(route = UserNavigationRoute) {
-        UserRoute()
+        MyPageRoute()
     }
 }
