+++ presenter/src/main/java/com/mashup/twotoo/presenter/ui/TwoTooAppState.kt
@@ -19,8 +19,8 @@ import com.mashup.twotoo.presenter.navigation.TopLevelDestination
 import com.mashup.twotoo.presenter.navigation.TopLevelDestination.Garden
 import com.mashup.twotoo.presenter.navigation.TopLevelDestination.Home
 import com.mashup.twotoo.presenter.navigation.TopLevelDestination.User
-import com.mashup.twotoo.presenter.user.navigation.UserNavigationRoute
-import com.mashup.twotoo.presenter.user.navigation.navigateToUser
+import com.mashup.twotoo.presenter.mypage.navigation.UserNavigationRoute
+import com.mashup.twotoo.presenter.mypage.navigation.navigateToUser
 
 @Composable
 fun rememberTwoTooAppState(
