+++ presenter/src/main/res/values/string.xml
@@ -3,7 +3,7 @@
     <string name="app_name">Twotoo</string>
     <string name="garden">Garden</string>
     <string name="home">Home</string>
-    <string name="user">User</string>
+    <string name="mypage">MyPage</string>
 
     <string name="home_nav_button">홈 네비게이션 버튼</string>
     <string name="garden_nav_button">가든 네비게이션 버튼</string>
